requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
webrtcvad>=2.0.10
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pyyaml>=6.0